moved assets folder to outside app
